pyproject.toml
[tool.poetry]
name = "coursework-7-drf"
version = "0.1.0"
description = ""
authors = ["Александра <aleksandrasilina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
black = "^24.8.0"
isort = "^5.13.2"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
eventlet = "^0.37.0"
redis = "^5.1.0"
requests = "^2.32.3"
django-celery-beat = "^2.7.0"
flake8 = "^7.1.1"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"
django-cors-headers = "^4.4.0"
coverage = "^7.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
